Updated schema

+++ app/src/main/java/com/android/stegware_app/api/schema/TimingSchema.java
@@ -1,12 +1,12 @@
 package com.android.stegware_app.api.schema;
 
 public class TimingSchema {
-    public int parseTime;
-    public int compileTime;
-    public int dynamicLoadingTime;
-    public int executionTime;
+    public double parseTime;
+    public double compileTime;
+    public double dynamicLoadingTime;
+    public double executionTime;
 
-    public TimingSchema(int parseTime, int compileTime, int dynamicLoadingTime, int executionTime) {
+    public TimingSchema(double parseTime, double compileTime, double dynamicLoadingTime, double executionTime) {
         this.parseTime = parseTime;
         this.compileTime = compileTime;
         this.dynamicLoadingTime = dynamicLoadingTime;
